Words-Server/requirements.txt
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Limiter==2.0.4
flask-mongoengine==1.0.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.2
html5lib==1.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
limits==2.1.0
MarkupSafe==2.0.1
mongoengine==0.23.1
psycopg2-binary==2.9.3
psycopg2-pool==1.1
pyls==0.1.6
pymongo==3.12.3
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
soupsieve==2.3.1
SQLAlchemy==1.4.29
urllib3==1.26.7
webencodings==0.5.1
Werkzeug==2.0.2
WTForms==3.0.1
